requirement-linux.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
bitsandbytes==0.49.2
build==1.4.0
CacheControl==0.14.4
certifi==2026.2.25
charset-normalizer==3.4.5
cleo==2.1.0
click==8.2.1
colorama==0.4.6
crashtest==0.4.1
cut-cross-entropy==25.1.1
dataclasses-json==0.6.7
datasets==4.3.0
diffusers==0.37.0
dill==0.4.0
distlib==0.4.0
docstring_parser==0.17.0
dulwich==1.1.0
fastjsonschema==2.21.2
filelock==3.25.2
findpython==0.7.1
frozenlist==1.8.0
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.4.2
hf_transfer==0.1.9
httpcore==1.0.9
httpx==0.28.1
idna==3.11
immutables==0.21
importlib_metadata==8.7.1
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.1.1
jaraco.functools==4.4.0
Jinja2==3.1.6
keyring==25.7.0
llm-blender==0.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
mergekit==0.1.4
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
msgspec==0.20.0
multidict==6.7.1
multiprocess==0.70.16
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.3.4
packaging==26.0
pandas==3.0.1
pbs-installer==2026.3.10
peft==0.18.1
pillow==12.1.1
pkginfo==1.12.1.2
platformdirs==4.9.4
poetry==2.3.2
poetry-core==2.3.1
propcache==0.4.1
protobuf==7.34.0
psutil==7.2.2
pyarrow==23.0.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.2
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-discovery==1.1.3
pywin32-ctypes==0.2.3
PyYAML==6.0.3
RapidFuzz==3.14.3
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
safetensors==0.5.3
scipy==1.17.1
sentencepiece==0.2.1
setuptools==82.0.1
setuptools-scm==9.2.2
shellingham==1.5.4
six==1.17.0
sympy==1.14.0
tokenizers==0.22.2
toml==0.10.2
tomlkit==0.14.0
torchao==0.15.0
tqdm==4.67.1
transformers==4.57.6
trl==0.24.0
trove-classifiers==2026.1.14.14
typeguard==4.5.1
typer==0.24.1
typer-slim==0.24.0
typing-inspect==0.9.0
typing_extensions==4.15.0
tyro==1.0.9
tzdata==2025.3
unsloth==2026.3.4
unsloth_zoo==2026.3.2
urllib3==2.6.3
virtualenv==21.2.0
wheel==0.46.3
xformers==0.0.33.post2
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
zstandard==0.25.0
gguf==0.18.0